Rename marquee block to video-marquee, fix layout and align to Figma spec
marquee is a reserved milo core block name, so a local block by that name
was silently shadowed by milo's own built-in marquee (that's why nothing
ever rendered). Renamed to video-marquee to avoid the collision, switched
decoration to mutate existing authored elements in place instead of
wiping and rebuilding the block, fixed the video column overflowing past
its grid track (missing min-width: 0), made the video full-bleed to the
right edge instead of centered, fixed video height being driven by
aspect-ratio against the wrong width, and swapped in the provided
play/mute icon assets. Headline and subcopy typography now match the
exact Figma spec values.
